# [Support] Validate number of arguments passed to formatv()