Change formatv() to validate that the number of arguments passed matches number of 
replacement fields in the format string, and that the replacement indices do not contain
holes.

To support cases where this cannot be guaranteed, introduce a formatv() variant that does not
do this validation (called formatvv()).

Fixed existing issues found by this validation.