@jurahul 2024-08-27
Hi all, this is related to https://discourse.llvm.org/t/adding-argument-count-validation-for-formatv/80876/1

I still have formatv() and formatvv() functions in the code, but only a handful instances. So, if this looks ok overall, I will go ahead and rename formatvv() to formatv_unchecked() (or if there are other better suggestions for that version). Also, let me know if it would make sense to split this into 2 changes, one NFC one that fixes formatv() current uses that are invalid, and then the next one that adds validation and changes the handful few that need to be to formatvv().