@jurahul 2024-08-27
> compile-time tests with clang for example

I can check. How do I run them?