@jurahul 2024-08-27
> > compile-time tests with clang for example
> 
> I can check. How do I run them?

I am wondering if running mlir-tblgen is a better benchmark, given that formatv() is widely used there (as opposed to clang), in case this measurement has to be done manually.