@joker-eph 2024-08-28
Having this enabled guarded by NDEBUG seems like a good option. There is already some things already like this isn't there? I kind of remember that using imbalanced `{` / `}` pair of something like that was checked only in assert builds?