@jurahul 2024-08-28
Yeah. For the unbalanced {{ I actually committed a change to fail in
release builds as well by printing an error message as formatv() output.
The only issue with debug only is that many folks build tablegen in release
mode, so those bad uses may not get flagged. But as long as the CI debug
builds use a debug build of tablegen, that would be ok. Let me work on
making this debug only.

On Tue, Aug 27, 2024 at 5:20 PM Mehdi Amini ***@***.***>
wrote:

> Having this enabled guarded by NDEBUG seems like a good option. There is
> already some things already like this isn't there? I kind of remember that
> using imbalanced { / } pair of something like that was checked only in
> assert builds?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/llvm/llvm-project/pull/105745#issuecomment-2313792375>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/APRMUB7B2AZJEODRN5COHSLZTUJTZAVCNFSM6AAAAABM7BSHH6VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDGMJTG44TEMZXGU>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>