@jurahul 2024-08-28
Looks like the CI builds are release but with LLVM_ENABLE_ASSERTION=ON, so this validation code will be enabled and potential bad usage will be flagged as well going forward as long as it's exercised during the build or test.