@jurahul 2024-08-28
> LGTM

Thanks. I'll split out an additional PR for the independent fixes, and then in this one I'll change ENABLE_VALIDATION back to 0 and also remove the `formatvv()` and use `formatv(false` instead so we just have one function name.