@jurahul 2024-08-29
The final version is good to go @joker-eph if you want to take another look. Nothing significant changed, just formatvv() is now gone and replaced with formatv(false,...).